# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /spend/pools:
    get:
      operationId: listSpendPools
      tags: [groups]
      summary: List authorized canonical budget pools
      parameters:
        - $ref: "#/components/parameters/RangeType"
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - $ref: "#/components/parameters/ViewScope"
        - $ref: "#/components/parameters/SpendSearch"
        - $ref: "#/components/parameters/SpendStatus"
        - $ref: "#/components/parameters/SpendWorkspace"
        - $ref: "#/components/parameters/SpendSort"
        - $ref: "#/components/parameters/SpendPage"
        - $ref: "#/components/parameters/SpendPageSize"
      responses:
        "200":
          description: Canonical pool page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpendTableResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /spend/groups:
    get:
      operationId: listSpendGroups
      tags: [groups]
      summary: List authorized workspace-qualified groups
      parameters:
        - $ref: "#/components/parameters/RangeType"
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - $ref: "#/components/parameters/ViewScope"
        - $ref: "#/components/parameters/SpendSearch"
        - $ref: "#/components/parameters/SpendStatus"
        - $ref: "#/components/parameters/SpendWorkspace"
        - $ref: "#/components/parameters/SpendSort"
        - $ref: "#/components/parameters/SpendPage"
        - $ref: "#/components/parameters/SpendPageSize"
      responses:
        "200":
          description: Group page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpendTableResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /spend/people:
    get:
      operationId: listSpendPeople
      tags: [groups]
      summary: List authorized workspace-qualified people usage
      parameters:
        - $ref: "#/components/parameters/RangeType"
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - $ref: "#/components/parameters/ViewScope"
        - $ref: "#/components/parameters/SpendSearch"
        - $ref: "#/components/parameters/SpendStatus"
        - $ref: "#/components/parameters/SpendWorkspace"
        - $ref: "#/components/parameters/SpendSort"
        - $ref: "#/components/parameters/SpendPage"
        - $ref: "#/components/parameters/SpendPageSize"
      responses:
        "200":
          description: People page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpendTableResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /spend/projects:
    get:
      operationId: listSpendProjects
      tags: [groups]
      summary: List authorized workspace-qualified project usage
      parameters:
        - $ref: "#/components/parameters/RangeType"
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - $ref: "#/components/parameters/ViewScope"
        - $ref: "#/components/parameters/SpendSearch"
        - $ref: "#/components/parameters/SpendStatus"
        - $ref: "#/components/parameters/SpendWorkspace"
        - $ref: "#/components/parameters/SpendSort"
        - $ref: "#/components/parameters/SpendPage"
        - $ref: "#/components/parameters/SpendPageSize"
      responses:
        "200":
          description: Project page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpendTableResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /spend/pools.csv:
    get:
      operationId: exportSpendPoolsCsv
      tags: [groups]
      summary: Export the authorized filtered canonical pool rows
      parameters:
        - $ref: "#/components/parameters/RangeType"
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - $ref: "#/components/parameters/ViewScope"
        - $ref: "#/components/parameters/SpendSearch"
        - $ref: "#/components/parameters/SpendStatus"
        - $ref: "#/components/parameters/SpendWorkspace"
        - $ref: "#/components/parameters/SpendSort"
      responses:
        "200":
          $ref: "#/components/responses/SpendCsv"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /spend/groups.csv:
    get:
      operationId: exportSpendGroupsCsv
      tags: [groups]
      summary: Export the authorized filtered group rows
      parameters:
        - $ref: "#/components/parameters/RangeType"
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - $ref: "#/components/parameters/ViewScope"
        - $ref: "#/components/parameters/SpendSearch"
        - $ref: "#/components/parameters/SpendStatus"
        - $ref: "#/components/parameters/SpendWorkspace"
        - $ref: "#/components/parameters/SpendSort"
      responses:
        "200":
          $ref: "#/components/responses/SpendCsv"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /spend/people.csv:
    get:
      operationId: exportSpendPeopleCsv
      tags: [groups]
      summary: Export the authorized filtered people rows
      parameters:
        - $ref: "#/components/parameters/RangeType"
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - $ref: "#/components/parameters/ViewScope"
        - $ref: "#/components/parameters/SpendSearch"
        - $ref: "#/components/parameters/SpendStatus"
        - $ref: "#/components/parameters/SpendWorkspace"
        - $ref: "#/components/parameters/SpendSort"
      responses:
        "200":
          $ref: "#/components/responses/SpendCsv"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /spend/projects.csv:
    get:
      operationId: exportSpendProjectsTableCsv
      tags: [groups]
      summary: Export the authorized filtered project rows
      parameters:
        - $ref: "#/components/parameters/RangeType"
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - $ref: "#/components/parameters/ViewScope"
        - $ref: "#/components/parameters/SpendSearch"
        - $ref: "#/components/parameters/SpendStatus"
        - $ref: "#/components/parameters/SpendWorkspace"
        - $ref: "#/components/parameters/SpendSort"
      responses:
        "200":
          $ref: "#/components/responses/SpendCsv"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  responses:
    SpendCsv:
      description: CSV containing exactly the authorized filtered rows in server sort order.
      headers:
        X-Filtered-Rows:
          schema: { type: integer, minimum: 0 }
        X-Total-Spend-Usd:
          schema: { type: number }
        X-Generation-Id:
          schema: { type: string }
        X-Usage-Range:
          description: UTC end-exclusive committed reporting interval.
          schema: { type: string }
      content:
        text/csv:
          schema: { type: string }
    Unauthorized:
      description: Authentication required — no valid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: >
        Authenticated but not authorized (neither account admin nor an enabled
        workspace admin, or attempting an account-admin-only action).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  parameters:
...
    RangeType:
      name: rangeType
      in: query
      required: false
      description: Date range for usage. billing = current billing cycle (default), full-term = rolling May 20, 2026 through today, mtd = month to date, ytd = year to date, custom requires startDate and endDate.
      schema:
        type: string
        enum: [billing, full-term, mtd, ytd, custom]
    ViewScope:
      name: viewScope
      in: query
      required: false
      description: Server-resolved presentation scope; managed excludes unrelated self-only grants.
      schema:
        type: string
        enum: [managed, my, all_authorized]
...
    SpendSearch:
      name: search
      in: query
      required: false
      schema:
        type: string
        maxLength: 200
    SpendStatus:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [all, over, attention, budgeted, unbudgeted, no_allocation, shared, explicit, inherited, no_limit, unavailable]
    SpendWorkspace:
      name: workspaceId
      in: query
      required: false
      description: Exact authorized workspace facet. Omit to include every workspace in the resolved scope.
      schema:
        type: string
        maxLength: 200
    SpendSort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: [status, spend_desc, spend_asc, name_asc, name_desc]
    SpendPage:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    SpendPageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    StartDate:
      name: startDate
      in: query
      required: false
      description: Inclusive UTC start date (YYYY-MM-DD), required when rangeType=custom
      schema:
        type: string
    EndDate:
      name: endDate
      in: query
      required: false
      description: Inclusive UTC end date (YYYY-MM-DD), required when rangeType=custom
      schema:
        type: string
...
  schemas:
...
    ReportingPeriod:
      type: object
      required: [start, endExclusive, timezone, label]
      properties:
        start: { type: string, format: date-time }
        endExclusive: { type: string, format: date-time }
        timezone: { type: string, const: UTC }
        label: { type: string }
    AccountingMetadata:
      type: object
      required: [generationId, costBasis, status, dataAsOf, directoryDataAsOf, stale, coverage, qualifications, limitObservation]
      properties:
        generationId: { type: string }
        costBasis:
          type: string
          enum: [allocation_eligible_committed]
        status:
          type: string
          enum: [complete, stale, partial, empty]
        dataAsOf: { type: [string, "null"], format: date-time }
        directoryDataAsOf: { type: [string, "null"], format: date-time }
        stale: { type: boolean }
        coverage:
          type: object
          required: [ratio, requestedDays, missingDays, failedWorkspaceDays]
          properties:
            ratio: { type: number, minimum: 0, maximum: 1 }
            requestedDays: { type: integer, minimum: 0 }
            missingDays:
              type: array
              items: { type: string }
            failedWorkspaceDays:
              type: array
              items: { type: string }
        qualifications:
          type: array
          items: { type: string }
        limitObservation:
          type: object
          required: [status, observedAt, lastSuccessfulAt, lastAttemptAt, refreshStartedAt, generation, error]
          properties:
            status:
              type: string
              enum: [complete, failed, unavailable, refreshing]
            observedAt: { type: [integer, "null"] }
            lastSuccessfulAt:
              type: [integer, "null"]
              description: Time of the last authoritative complete observation, retained through later failures.
            lastAttemptAt: { type: [integer, "null"] }
            refreshStartedAt: { type: [integer, "null"] }
            generation:
              type: [string, "null"]
              description: Durable identity of the last successful set of limit values.
            error: { type: [string, "null"] }
    DashboardScope:
      type: object
      required: [viewScope, label, workspaceIds, groupIds, isPersonal]
      properties:
        viewScope:
          type: string
          enum: [managed, my, all_authorized]
        label: { type: string }
        workspaceIds:
          type: array
          items: { type: string }
        groupIds:
          type: array
          items: { type: string }
        isPersonal: { type: boolean }
...
    SpendTableRow:
      type: object
      required: [id, kind, name, workspaceId, workspaceName, spendUsd, agentSpendUsd, otherServicesUsd, allocationUsd, remainingUsd, percentUsed, status, memberCount, ownerName, limitState, limitObservationStatus, sharedPool]
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [pool, group, person, project, unattributed, reconciliation]
        name: { type: string }
        workspaceId: { type: [string, "null"] }
        workspaceName: { type: [string, "null"] }
        spendUsd: { type: number }
        agentSpendUsd: { type: number }
        otherServicesUsd: { type: number }
        allocationUsd: { type: [number, "null"] }
        remainingUsd: { type: [number, "null"] }
        percentUsed: { type: [number, "null"] }
        currentCycleAgentSpendUsd:
          type: [number, "null"]
          description: Agent spend in the current billing cycle; populated for People independently of the selected reporting range.
        currentCycleRemainingUsd:
          type: [number, "null"]
          description: Current monthly limit minus current-cycle Agent spend; never selected-range spend.
        currentCyclePercentUsed:
          type: [number, "null"]
          description: Current-cycle Agent spend as a percentage of the current monthly limit.
        status: { type: string }
        memberCount: { type: [integer, "null"] }
        ownerName: { type: [string, "null"] }
        limitState:
          type: string
          enum: [not_applicable, explicit, inherited, no_limit, unavailable]
        limitObservationStatus:
          type: string
          enum: [not_applicable, complete, failed, unavailable, refreshing]
          description: Current durable observation state; last successful values may remain visible during refresh or after failure.
        sharedPool: { type: boolean }
    SpendTableResponse:
      type: object
      required: [view, scope, period, rows, page, pageSize, totalRows, filteredRows, totals, facets, metadata]
      properties:
        view:
          type: string
          enum: [pools, groups, people, projects]
        scope:
          $ref: "#/components/schemas/DashboardScope"
        period:
          $ref: "#/components/schemas/ReportingPeriod"
        rows:
          type: array
          items:
            $ref: "#/components/schemas/SpendTableRow"
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        totalRows: { type: integer, minimum: 0 }
        filteredRows: { type: integer, minimum: 0 }
        totals:
          type: object
          required: [spendUsd, agentSpendUsd, otherServicesUsd, allocationUsd, internalExcludedUsd, unbudgetedUsd, unattributedUsd, reconciliationUsd]
          properties:
            spendUsd: { type: number }
            agentSpendUsd: { type: number }
            otherServicesUsd: { type: number }
            allocationUsd: { type: number }
            internalExcludedUsd: { type: number }
            unbudgetedUsd: { type: number }
            unattributedUsd: { type: number }
            reconciliationUsd: { type: number }
        facets:
          type: object
          required: [statuses, workspaces]
          properties:
            statuses:
              type: object
              additionalProperties: { type: integer, minimum: 0 }
            workspaces:
              type: array
              items:
                type: object
                required: [id, name, count]
                properties:
                  id: { type: string }
                  name: { type: string }
                  count: { type: integer, minimum: 0 }
        metadata:
          $ref: "#/components/schemas/AccountingMetadata"
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...