# lib/api-zod/src/generated/types/spendTableRow.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SpendTableRowKind } from './spendTableRowKind';
import type { SpendTableRowLimitObservationStatus } from './spendTableRowLimitObservationStatus';
import type { SpendTableRowLimitState } from './spendTableRowLimitState';

export interface SpendTableRow {
  id: string;
  kind: SpendTableRowKind;
  name: string;
  /** @nullable */
  workspaceId: string | null;
  /** @nullable */
  workspaceName: string | null;
  spendUsd: number;
  agentSpendUsd: number;
  otherServicesUsd: number;
  /** @nullable */
  allocationUsd: number | null;
  /** @nullable */
  remainingUsd: number | null;
  /** @nullable */
  percentUsed: number | null;
  /**
     * Agent spend in the current billing cycle; populated for People independently of the selected reporting range.
     * @nullable
     */
  currentCycleAgentSpendUsd?: number | null;
  /**
     * Current monthly limit minus current-cycle Agent spend; never selected-range spend.
     * @nullable
     */
  currentCycleRemainingUsd?: number | null;
  /**
     * Current-cycle Agent spend as a percentage of the current monthly limit.
     * @nullable
     */
  currentCyclePercentUsed?: number | null;
  status: string;
  /** @nullable */
  memberCount: number | null;
  /** @nullable */
  ownerName: string | null;
  limitState: SpendTableRowLimitState;
  /** Current durable observation state; last successful values may remain visible during refresh or after failure. */
  limitObservationStatus: SpendTableRowLimitObservationStatus;
  sharedPool: boolean;
}